refactor(#63): 프론트 백엔드 연결 리팩토링 및 채팅 도메인 관련 프론트 백엔드 수정

+++ backend/src/main/java/com/example/readys7project/domain/project/dto/request/ProjectRequestDto.java
@@ -20,10 +20,10 @@ public record ProjectRequestDto(
     Long categoryId,
 
     @NotNull(message = "최저 예산은 필수입니다.")
-    Integer minBudget,
+    Long minBudget,
 
     @NotNull(message = "최대 예산은 필수입니다.")
-    Integer maxBudget,
+    Long maxBudget,
 
     @NotNull(message = "기간은 필수입니다.")
     Integer duration,
